Improving map JS security
